Split sharing tests into files per scenario.
